+++ lib/ZMQ/FFI/ZMQ3/Context.pm
@@ -114,6 +114,7 @@ sub socket {
         ctx          => $self,
         type         => $type,
         soname       => $self->soname,
+        autoclean    => $self->autoclean,
     );
 }
 
@@ -155,6 +156,8 @@ sub destroy {
 
 sub DEMOLISH {
     my ($self) = @_;
+
+    return if not $self->autoclean;
 
     unless ($self->_ctx == -1) {
         $self->destroy();
